Refactoring for authentication - Prometheus added

+++ src/main/java/com/fitfinance/service/FinanceService.java
@@ -2,7 +2,6 @@ package com.fitfinance.service;
 
 import com.fitfinance.domain.Finance;
 import com.fitfinance.domain.User;
-import com.fitfinance.exception.EmailAlreadyExistsException;
 import com.fitfinance.exception.NotFoundException;
 import com.fitfinance.repository.FinanceRepository;
 import com.fitfinance.repository.UserRepository;
@@ -16,13 +15,18 @@ import java.util.List;
 @RequiredArgsConstructor
 public class FinanceService {
     private final FinanceRepository financeRepository;
+    private final UserRepository userRepository;
 
     public List<Finance> getAllFinances() {
         return financeRepository.findAll();
     }
 
     public Finance findById(Long id) {
-        return financeRepository.findById(id).orElseThrow(() -> new NotFoundException("User not found"));
+        return financeRepository.findById(id).orElseThrow(() -> new NotFoundException("Finance not found"));
+    }
+
+    public List<Finance> findByUserId(Long userId) {
+        return financeRepository.findByUserId(userId);
     }
 
     @Transactional
@@ -31,13 +35,22 @@ public class FinanceService {
     }
 
     public void updateFinance(Finance financeToUpdate) {
-        var foundFinance = findById(financeToUpdate.getId());
+        Finance foundFinance = findById(financeToUpdate.getId());
 
-        financeRepository.save(foundFinance);
+        if (!financeToUpdate.getUser().getId().equals(foundFinance.getUser().getId())) {
+            throw new SecurityException("User does not have permission");
+        }
+
+        financeRepository.save(financeToUpdate);
     }
 
-    public void deleteFinance(Long id) {
+    public void deleteFinance(Long id, User user) {
         var finance = findById(id);
+
+        if (!finance.getUser().getId().equals(user.getId())) {
+            throw new SecurityException("User does not have permission");
+        }
+
         financeRepository.delete(finance);
     }
 }
